added code for updating profile
