+++ pages/profile.html
@@ -42,7 +42,7 @@
               <div class="text-center">
                 <div class="text-center">
                   <!-- profile image -->
-                  <button class="btn btn-primary" type="button">Edit Profile</button>
+                  <button class="btn btn-primary" type="button" href="/pages/updateProfile.html">Edit Profile</button>
                   <div class="profile-img clearfix profile-img-lg profile-img-center mt-n5 mb-3" id="profileImg">
                     <a href="#"><img class="profile-img rounded border border-white border-3" src="https://xsgames.co/randomusers/avatar.php?g=male" alt="Image of the profile owner"></a>
                   </div>
@@ -428,7 +428,7 @@
     <!-- <script src="/src/js/index.mjs" type="module"></script> -->
     <script src="/src/js/auth/check.js"></script>
     <script src="/src/js/auth/logout.js"></script>
-    <script src="/src/js/profile/userprofile.mjs" type="module"></script>
+    <script src="/src/js/profile/getProfile.mjs" type="module"></script>
     <script src="/src/js/profile/userposts.mjs" type="module"></script>
     <!-- <script src="/src/js/profile/updateprofile.mjs"></script> -->
