+++ pages/updateProfile.html
@@ -108,8 +108,6 @@
     <script src="./src/js/auth/check.js"></script>
     <script src="src/js/auth/logout.js"></script>
     <script src="src/js/index.mjs" type="module"></script>
-    <script src="src/js/posts/createPost.mjs" type="module"></script>
-    <script src="src/js/posts/posts.mjs" type="module"></script>
-    <!-- <script src="src/js/components/deletePost.mjs" type="module"></script> -->
+    <script src="src/js/profile/getProfile.mjs" type="module"></script>
   </body>
 </html>
