+++ src/js/profile/index.mjs
@@ -1,0 +1,2 @@
+export * from "./getProfile.mjs";
+export * from "./updateprofile.mjs";
