+++ src/js/profile/updateprofile.mjs
@@ -1,41 +1,19 @@
-import { getProfile, updateProfile } from "../api/profile/index.mjs";
-import { load } from "../storage/storage.mjs";
+import { API_SOCIAL_URL } from "../constants.mjs";
+import { authFetch } from "../authFetch.mjs";
 
-export async function setUpdateProfileListener() {
-  const form = document.querySelector("#updateProfileForm");
+const action = "/profiles";
+const method = "put";
 
- 
-
-  if (form) {
-    const { name, email } = load("profile")
-    form.name.value = name;
-    form.email.value = email;
-
-    const button = form.querySelector("button");
-    button.disabled = true;
-
-    const profile = await getProfile(name);
-
+export async function uppdateProfile(profileData) {
+    if (!profileData.name) {
+        throw new Error("Check your form");
+    }
+    const updateProfileURL = `${API_SOCIAL_URL}${action}/${profileData.name}/media`;
     
-    form.banner.value = profile.banner;
-    form.avatar.value = profile.avatar
+    const response = await authFetch(updateProfileURL, {
+        method,
+        body: JSON.stringify(profileData)
+    })
 
-    button.disabled = false;
-    
-    form.addEventListener("submit", (event) => {
-      event.preventDefault();
-      const form = event.target;
-      const formData = new FormData(form);
-      const profile = Object.fromEntries(formData.entries())
-      
-      profile.name = name;
-      profile.email = email;
-      console.log('success');
-        
-      //calling the function to send to API
-      updateProfile(profile)
-    });
-  }
-
-
+    return await response.json();
 }
